I18n: Defer translation of categories until use
The categories are used to sort the applications and are hence stored in
the `.app` files for comparison; the saved values are in English which
means that if the categories are translated too soon then the
comparisons will fail and the apps will not appear in the correct
categories. Resolve this problem by deferring the translation of these
names until after the comparisons have been made.