Fix creating localised .desktop files
Given that the binary `update-app-dir` is not installing the i18n module, it
seems to not be prepared to handle non-ascii strings. However, `AppData` would
previously return a localised set of apps without checking whether the i18n
was registered and installed.

This is a temporary fix and `update-app-dir` should be prepared to create
localised .desktop app files.