Qt i18n: Add support for reading localised ".app" files
Add awareness of localised `.app` files when reading them for display as
an icon grid. Implements a fallback system so should some data be
missing from the localised `.app` file then it will supplement it with
that from the default set.

Note: A limitation with the implementation is that boolean types will
not fall back.