i18n: Properly merge translated `.app` files into English
If the translated `.app` files do not contain the `launch_command` field
then the app merge would fail. Resolve this by taking the English
version and updating it with the unparsed i18n version.

Also modify the `update-app-dir` script to insert translated versions of
fields supported in the `auto_*.desktop` files.