Priority: Add a priority to apps on downloading
Add the `priority` field to apps upon download when they do not provide
such field themselves. Ideally, the knowledge of the priority would be
stored entirely in the API and there would be no need for this, however
until those changes are made, assign them on download.

Also lower the priority of the `terminal` app to give precedence for
Terminal Quest.